Refactor component testing page
