--- nugurang-front/pages/comp_test/article_box_with_accordion_test.js
@@ -1,19 +1,0 @@
-import React from 'react';
-
-import ArticleBoxWithAccordion from '../../components/ArticleBoxWithAccordion';
-
-export default function CompTest() {
-
-const article = {
-  title: "Article title",
-  subtitle: "Article subtitle",
-  image: "/static/images/sample_1.jpg",
-  content: "Article content using accordion",
-};
-
-  return (
-	<div>
-	  <ArticleBoxWithAccordion article={article} />
-	</div>
-  );
-}
