--- nugurang-front/pages/comp_test/content_paper_test.js
@@ -1,12 +1,0 @@
-import React from 'react';
-
-import ContentPaper from '../../components/ContentPaper';
-
-export default function CompTest() {
-
-  return (
-	<div>
-	  <ContentPaper />
-	</div>
-  );
-}
