--- nugurang-front/pages/comp_test/page_title_box_test.js
@@ -1,14 +1,0 @@
-import React from 'react';
-
-import PageTitleBox from '../../components/PageTitleBox';
-
-export default function CompTest() {
-
-  let title = 'Back';
-  
-  return (
-	<div>
-	  <PageTitleBox title={title} />
-	</div>
-  );
-}
