--- nugurang-front/pages/comp_test/page_title_box_without_back_button_test.js
@@ -1,14 +1,0 @@
-import React from 'react';
-
-import PageTitleBoxWithoutBackButton from '../../components/PageTitleBoxWithoutBackButton';
-
-export default function CompTest() {
-
-  let title = 'Back';
-  
-  return (
-	<div>
-	  <PageTitleBoxWithoutBackButton title={title} />
-	</div>
-  );
-}
