--- nugurang-front/pages/comp_test/section_title_box_test.js
@@ -1,15 +1,0 @@
-import React from 'react';
-
-import AssignmentIcon from '@material-ui/icons/Assignment';
-import SectionTitleBox from '../../components/SectionTitleBox';
-
-export default function CompTest() {
-	
-  let title = 'Title';
-
-  return (
-	<div>
-	  <SectionTitleBox title={title} icon={<AssignmentIcon />} />
-	</div>
-  );
-}
