--- nugurang-front/pages/comp_test/section_title_box_with_text_field_test.js
@@ -1,15 +1,0 @@
-import React from 'react';
-
-import AssignmentIcon from '@material-ui/icons/Assignment';
-import SectionTitleBoxWithTextField from '../../components/SectionTitleBoxWithTextField';
-
-export default function CompTest() {
-	
-  let title = 'Title';
-
-  return (
-	<div>
-	  <SectionTitleBoxWithTextField title={title} icon={<AssignmentIcon />} />
-	</div>
-  );
-}
