--- nugurang-front/pages/comp_test/universal_button_test.js
@@ -1,15 +1,0 @@
-import React from 'react';
-
-import AssignmentIcon from '@material-ui/icons/Assignment';
-import UniversalButton from '../../components/UniversalButton';
-
-export default function CompTest() {
-
-let label = "UniversalButton";
-
-  return (
-	<div>
-	  <UniversalButton label={label}/>
-	</div>
-  );
-}
